refactor: moved MovieFavorite from MovieCard into ui components
